// src/fabric/api/FabricApiClient.ts
export type RuntimeAttribute = RuntimeType.DotNet | RuntimeType.Python; // workload also has 'NOTASSIGNED', but we'll query user for value if not set
export type InputTypeAttribute = InputType.Http | InputType.EventStream;

export type ArtifactAttributes = {
    runtime?: RuntimeAttribute;
    // 'inputType'?: InputTypeAttribute
};
/**
 * IArtifact as seen in api responses
 */
export interface IArtifact {
    id: string;
    type: string;
    displayName: string;
    description: string | undefined;
    workspaceId: string;
    attributes?: ArtifactAttributes;

    /** Represent the Fabric environment this item exists in, like "DAILY" or "PROD".
     *
     * It's debatable whether this should be here as it's not part of the API response,
     * however we use IArtifact throughout the codebase and it's useful to have it here.
     */
    fabricEnvironment: string;
}